refactor(fsd): segment export and settings slices
