+++ src/features/settings/model/hooks/useConfigFormState.ts
@@ -9,7 +9,7 @@ import type { ConfigState } from "@/shared/config";
 import * as React from "react";
 import { useForm, useWatch } from "react-hook-form";
 
-import type { ConfigFormProps } from "@/features/settings/components/ConfigForm";
+import type { ConfigFormProps } from "@/features/settings/ui/components/ConfigForm";
 
 export interface UseConfigFormStateOptions {
   config: ConfigState;
