+++ src/features/settings/ui/components/ConfigForm.tsx
@@ -10,7 +10,7 @@ import type * as React from "react";
 import { useId } from "react";
 import { AiOutlineDown, AiOutlineEye, AiOutlinePlayCircle, AiOutlineTool, AiOutlineUser } from "react-icons/ai";
 
-import { SettingsRow, SettingsSection } from "@/features/settings/components/SettingsSection";
+import { SettingsRow, SettingsSection } from "@/features/settings/ui/components/SettingsSection";
 import { Button } from "@/shared/ui/button";
 import { Slider, Switch } from "@/shared/ui/forms";
