+++ steiger.config.ts
@@ -13,15 +13,9 @@ export default defineConfig([
     },
   },
   {
-    files: ["./src/entities/study-progress/**", "./src/features/deck-editor/**"],
+    files: ["./src/entities/study-progress/**", "./src/features/deck-editor/**", "./src/features/export/**"],
     rules: {
       "fsd/insignificant-slice": "off",
-    },
-  },
-  {
-    files: ["./src/features/export/**", "./src/features/settings/**"],
-    rules: {
-      "fsd/no-segmentless-slices": "off",
     },
   },
 ]);
